refactor(site): extract AgentsLayout from AgentsPageView
Rename AgentsPageView to AgentsLayout to follow the convention used
by HealthLayout, TemplateLayout, and other layout components. Move
the scrollbar gutter lock effect from AgentsPage into the layout
so all viewport-level concerns live in one place.
